refactor(archive): move semantic facts out of lib
