+++ devtools/build_topology_projection.py
@@ -96,9 +96,9 @@ LIB_PREFIX_TO_SUBPACKAGE = {
     "message_": "lib/message/",
     "messages": "lib/message/",
     "conversation_": "lib/conversation/",
-    "semantic_fact_": "lib/semantic/",
-    "semantic_facts": "lib/semantic/facts.py",
-    "content_projection": "lib/semantic/",
+    "semantic_fact_": "archive/semantic/",
+    "semantic_facts": "archive/semantic/facts.py",
+    "content_projection": "archive/semantic/",
     "branch_type": "lib/conversation/",
     "threads": "lib/conversation/",
     "neighbor_candidates": "lib/conversation/",
@@ -210,7 +210,7 @@ TARGET_TO_OWNER = [
     ("polylogue/archive/action_event/", "archive-action-event"),
     ("polylogue/lib/message/", "lib-message"),
     ("polylogue/lib/conversation/", "lib-conversation"),
-    ("polylogue/lib/semantic/", "lib-semantic"),
+    ("polylogue/archive/semantic/", "archive-semantic"),
     ("polylogue/archive/phase/", "archive-phase"),
     ("polylogue/archive/projection/", "archive-projection"),
     ("polylogue/archive/provider/", "archive-provider"),
