+++ docs/topology-status.md
@@ -20,10 +20,11 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-provider | — | 3 | 3 | 0 | 0 |
 | archive-query | archive query semantics | 17 | 17 | 0 | 0 |
 | archive-raw-payload | — | 5 | 5 | 0 | 0 |
+| archive-semantic | — | 5 | 5 | 0 | 0 |
 
 ### Summary
 
-- **Stable** (no move scoped): 612
+- **Stable** (no move scoped): 607
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (lib-root + storage-root): 29
 - **TBD** (cell needs explicit assignment): 0
