+++ polylogue/cli/query_contracts.py
@@ -10,9 +10,9 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Literal, TypeAlias
 
 from polylogue.archive.query.spec import ConversationQuerySpec
+from polylogue.archive.semantic.content_projection import ContentProjectionSpec
 from polylogue.lib.message.roles import MessageRoleFilter, normalize_message_roles
 from polylogue.lib.roles import Role
-from polylogue.lib.semantic.content_projection import ContentProjectionSpec
 
 if TYPE_CHECKING:
     from polylogue.lib.models import Conversation, ConversationSummary
