+++ polylogue/lib/__init__.py
@@ -10,20 +10,20 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from polylogue.archive.attachment.models import Attachment
     from polylogue.archive.projection.projections import ConversationProjection
+    from polylogue.archive.semantic.content_projection import ContentKind, ContentProjectionSpec
     from polylogue.lib.conversation.branch_type import BranchType
     from polylogue.lib.conversation.models import Conversation
     from polylogue.lib.message.messages import MessageCollection
     from polylogue.lib.message.models import DialoguePair, Message
     from polylogue.lib.roles import Role
-    from polylogue.lib.semantic.content_projection import ContentKind, ContentProjectionSpec
 
 
 def __getattr__(name: str) -> object:
     lazy_exports = {
         "Attachment": ("polylogue.archive.attachment.models", "Attachment"),
         "BranchType": ("polylogue.lib.conversation.branch_type", "BranchType"),
-        "ContentKind": ("polylogue.lib.semantic.content_projection", "ContentKind"),
-        "ContentProjectionSpec": ("polylogue.lib.semantic.content_projection", "ContentProjectionSpec"),
+        "ContentKind": ("polylogue.archive.semantic.content_projection", "ContentKind"),
+        "ContentProjectionSpec": ("polylogue.archive.semantic.content_projection", "ContentProjectionSpec"),
         "Conversation": ("polylogue.lib.conversation.models", "Conversation"),
         "ConversationProjection": ("polylogue.archive.projection.projections", "ConversationProjection"),
         "DialoguePair": ("polylogue.lib.message.models", "DialoguePair"),
