+++ polylogue/lib/conversation/attribution.py
@@ -8,13 +8,13 @@ from pathlib import Path, PurePosixPath
 from typing import TYPE_CHECKING
 
 from polylogue.archive.action_event.action_events import ActionEvent
+from polylogue.archive.semantic.facts import ConversationSemanticFacts, build_conversation_semantic_facts
 from polylogue.lib.repo_identity import (
     normalize_repo_name,
     normalize_repo_names,
     normalize_repo_path,
     normalize_repo_paths,
 )
-from polylogue.lib.semantic.facts import ConversationSemanticFacts, build_conversation_semantic_facts
 from polylogue.lib.viewport.tools import looks_like_path_candidate
 
 if TYPE_CHECKING:
