+++ polylogue/lib/session/models.py
@@ -7,6 +7,7 @@ from dataclasses import dataclass
 from datetime import date, datetime
 
 from polylogue.archive.phase.extraction import SessionPhase
+from polylogue.archive.semantic.facts import ConversationSemanticFacts
 from polylogue.lib.conversation.attribution import ConversationAttribution
 from polylogue.lib.conversation.extraction import WorkEvent
 from polylogue.lib.payload_coercion import (
@@ -21,7 +22,6 @@ from polylogue.lib.payload_coercion import (
     string_sequence,
 )
 from polylogue.lib.repo_identity import normalize_repo_names, normalize_repo_paths
-from polylogue.lib.semantic.facts import ConversationSemanticFacts
 from polylogue.lib.session.documents import (
     SessionPhaseDocument,
     SessionProfileDocument,
