+++ polylogue/surfaces/payloads.py
@@ -12,8 +12,8 @@ from typing import TYPE_CHECKING, Literal, NotRequired
 from pydantic import BaseModel, ConfigDict, Field
 from typing_extensions import TypedDict
 
+from polylogue.archive.semantic.content_projection import ContentProjectionSpec
 from polylogue.lib.json import JSONDocument, JSONValue, require_json_document
-from polylogue.lib.semantic.content_projection import ContentProjectionSpec
 
 if TYPE_CHECKING:
     from collections.abc import Container
