+++ tests/infra/semantic_facts.py
@@ -11,9 +11,9 @@ from collections import Counter
 from collections.abc import Iterable, Sequence
 from dataclasses import dataclass
 
+from polylogue.archive.semantic.facts import build_conversation_semantic_facts
 from polylogue.lib.json import JSONDocument, json_document_list
 from polylogue.lib.models import Conversation
-from polylogue.lib.semantic.facts import build_conversation_semantic_facts
 from polylogue.storage.runtime import AttachmentRecord, ConversationRecord, MessageRecord
