+++ tests/unit/core/test_content_projection.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
 from polylogue.archive.attachment.models import Attachment
-from polylogue.lib.roles import Role
-from polylogue.lib.semantic.content_projection import (
+from polylogue.archive.semantic.content_projection import (
     ContentProjectionSpec,
     coerce_content_projection_spec,
     project_message_content,
 )
+from polylogue.lib.roles import Role
 from tests.infra.builders import make_conv, make_msg
