+++ tests/unit/core/test_semantic_facts.py
@@ -9,17 +9,17 @@ import pytest
 
 from polylogue.archive.phase.extraction import SessionPhase
 from polylogue.archive.phase.extraction import extract_phases as phase_extract_phases
-from polylogue.lib.conversation import extraction as work_event_extraction
-from polylogue.lib.message.messages import MessageCollection
-from polylogue.lib.models import Conversation as ConversationModel
-from polylogue.lib.models import ConversationSummary
-from polylogue.lib.pricing import harmonize_session_cost
-from polylogue.lib.semantic.facts import (
+from polylogue.archive.semantic.facts import (
     ConversationSemanticFacts,
     build_conversation_semantic_facts,
     build_mcp_summary_semantic_facts,
     build_projection_semantic_facts,
 )
+from polylogue.lib.conversation import extraction as work_event_extraction
+from polylogue.lib.message.messages import MessageCollection
+from polylogue.lib.models import Conversation as ConversationModel
+from polylogue.lib.models import ConversationSummary
+from polylogue.lib.pricing import harmonize_session_cost
 from polylogue.lib.session import runtime as session_profile_runtime
 from polylogue.lib.session.session_profile import build_session_profile
 from polylogue.storage.archive_views import ConversationRenderProjection
