+++ tests/unit/mcp/test_server_surfaces.py
@@ -10,9 +10,9 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from polylogue.archive.semantic.content_projection import ContentProjectionSpec
 from polylogue.lib.models import Conversation, ConversationSummary
 from polylogue.lib.roles import Role
-from polylogue.lib.semantic.content_projection import ContentProjectionSpec
 from polylogue.types import ConversationId, Provider
 from tests.infra.builders import make_conv, make_msg
 from tests.infra.mcp import (
